+++ library/src/schemas/index.ts
@@ -38,6 +38,7 @@ export * from './symbol/index.ts';
 export * from './tuple/index.ts';
 export * from './tupleWithRest/index.ts';
 export * from './undefined/index.ts';
+export * from './undefinedable/index.ts';
 export * from './union/index.ts';
 export * from './unknown/index.ts';
 export * from './variant/index.ts';