+++ library/src/schemas/looseObject/looseObject.test-d.ts
@@ -11,6 +11,7 @@ import {
   type StringIssue,
   type StringSchema,
 } from '../string/index.ts';
+import type { UndefinedableSchema } from '../undefinedable/index.ts';
 import { looseObject, type LooseObjectSchema } from './looseObject.ts';
 import type { LooseObjectIssue } from './types.ts';
 
@@ -46,6 +47,7 @@ describe('looseObject', () => {
         key3: NullishSchema<StringSchema<undefined>, never>;
         key4: ObjectSchema<{ key: NumberSchema<undefined> }, never>;
         key5: SchemaWithPipe<[StringSchema<undefined>, ReadonlyAction<string>]>;
+        key6: UndefinedableSchema<StringSchema<undefined>, 'bar'>;
       },
       undefined
     >;
@@ -58,6 +60,7 @@ describe('looseObject', () => {
           key3?: string | null | undefined;
           key4: { key: number };
           key5: string;
+          key6: string | undefined;
         } & { [key: string]: unknown }
       >();
     });
@@ -70,6 +73,7 @@ describe('looseObject', () => {
           key3?: string | null | undefined;
           key4: { key: number };
           readonly key5: string;
+          key6: string;
         } & { [key: string]: unknown }
       >();
     });