+++ library/src/schemas/looseObject/looseObjectAsync.test-d.ts
@@ -11,6 +11,7 @@ import {
   type StringIssue,
   type StringSchema,
 } from '../string/index.ts';
+import type { UndefinedableSchema } from '../undefinedable/index.ts';
 import {
   looseObjectAsync,
   type LooseObjectSchemaAsync,
@@ -51,6 +52,7 @@ describe('looseObjectAsync', () => {
         key3: NullishSchema<StringSchema<undefined>, never>;
         key4: ObjectSchema<{ key: NumberSchema<undefined> }, never>;
         key5: SchemaWithPipe<[StringSchema<undefined>, ReadonlyAction<string>]>;
+        key6: UndefinedableSchema<StringSchema<undefined>, 'bar'>;
       },
       undefined
     >;
@@ -63,6 +65,7 @@ describe('looseObjectAsync', () => {
           key3?: string | null | undefined;
           key4: { key: number };
           key5: string;
+          key6: string | undefined;
         } & { [key: string]: unknown }
       >();
     });
@@ -75,6 +78,7 @@ describe('looseObjectAsync', () => {
           key3?: string | null | undefined;
           key4: { key: number };
           readonly key5: string;
+          key6: string;
         } & { [key: string]: unknown }
       >();
     });