+++ library/src/schemas/objectWithRest/objectWithRest.test-d.ts
@@ -16,6 +16,7 @@ import {
   type StringIssue,
   type StringSchema,
 } from '../string/index.ts';
+import type { UndefinedableSchema } from '../undefinedable/index.ts';
 import { objectWithRest, type ObjectWithRestSchema } from './objectWithRest.ts';
 import type { ObjectWithRestIssue } from './types.ts';
 
@@ -55,6 +56,7 @@ describe('objectWithRest', () => {
         key3: NullishSchema<StringSchema<undefined>, never>;
         key4: ObjectSchema<{ key: NumberSchema<undefined> }, never>;
         key5: SchemaWithPipe<[StringSchema<undefined>, ReadonlyAction<string>]>;
+        key6: UndefinedableSchema<StringSchema<undefined>, 'bar'>;
       },
       BooleanSchema<undefined>,
       undefined
@@ -68,6 +70,7 @@ describe('objectWithRest', () => {
           key3?: string | null | undefined;
           key4: { key: number };
           key5: string;
+          key6: string | undefined;
         } & { [key: string]: boolean }
       >();
     });
@@ -80,6 +83,7 @@ describe('objectWithRest', () => {
           key3?: string | null | undefined;
           key4: { key: number };
           readonly key5: string;
+          key6: string;
         } & { [key: string]: boolean }
       >();
     });