+++ library/src/schemas/strictObject/strictObject.test-d.ts
@@ -11,6 +11,7 @@ import {
   type StringIssue,
   type StringSchema,
 } from '../string/index.ts';
+import type { UndefinedableSchema } from '../undefinedable/index.ts';
 import { strictObject, type StrictObjectSchema } from './strictObject.ts';
 import type { StrictObjectIssue } from './types.ts';
 
@@ -46,6 +47,7 @@ describe('strictObject', () => {
         key3: NullishSchema<StringSchema<undefined>, never>;
         key4: ObjectSchema<{ key: NumberSchema<undefined> }, never>;
         key5: SchemaWithPipe<[StringSchema<undefined>, ReadonlyAction<string>]>;
+        key6: UndefinedableSchema<StringSchema<undefined>, 'bar'>;
       },
       undefined
     >;
@@ -57,6 +59,7 @@ describe('strictObject', () => {
         key3?: string | null | undefined;
         key4: { key: number };
         key5: string;
+        key6: string | undefined;
       }>();
     });
 
@@ -67,6 +70,7 @@ describe('strictObject', () => {
         key3?: string | null | undefined;
         key4: { key: number };
         readonly key5: string;
+        key6: string;
       }>();
     });