+++ library/src/schemas/undefinedable/undefinedable.ts
@@ -1,0 +1,123 @@
+import { getDefault } from '../../methods/index.ts';
+import type {
+  BaseIssue,
+  BaseSchema,
+  Dataset,
+  Default,
+  InferInput,
+  InferIssue,
+} from '../../types/index.ts';
+import type { InferUndefinedableOutput } from './types.ts';
+
+/**
+ * Undefinedable schema type.
+ */
+export interface UndefinedableSchema<
+  TWrapped extends BaseSchema<unknown, unknown, BaseIssue<unknown>>,
+  TDefault extends Default<TWrapped, undefined>,
+> extends BaseSchema<
+    InferInput<TWrapped> | undefined,
+    InferUndefinedableOutput<TWrapped, TDefault>,
+    InferIssue<TWrapped>
+  > {
+  /**
+   * The schema type.
+   */
+  readonly type: 'undefinedable';
+  /**
+   * The schema reference.
+   */
+  readonly reference: typeof undefinedable;
+  /**
+   * The expected property.
+   */
+  readonly expects: `(${TWrapped['expects']} | undefined)`;
+  /**
+   * The wrapped schema.
+   */
+  readonly wrapped: TWrapped;
+  /**
+   * The default value.
+   */
+  readonly default: TDefault;
+}
+
+/**
+ * Creates a undefinedable schema.
+ *
+ * @param wrapped The wrapped schema.
+ *
+ * @returns A undefinedable schema.
+ */
+export function undefinedable<
+  const TWrapped extends BaseSchema<unknown, unknown, BaseIssue<unknown>>,
+>(wrapped: TWrapped): UndefinedableSchema<TWrapped, never>;
+
+/**
+ * Creates a undefinedable schema.
+ *
+ * @param wrapped The wrapped schema.
+ * @param default_ The default value.
+ *
+ * @returns A undefinedable schema.
+ */
+export function undefinedable<
+  const TWrapped extends BaseSchema<unknown, unknown, BaseIssue<unknown>>,
+  const TDefault extends Default<TWrapped, undefined>,
+>(
+  wrapped: TWrapped,
+  default_: TDefault
+): UndefinedableSchema<TWrapped, TDefault>;
+
+export function undefinedable(
+  wrapped: BaseSchema<unknown, unknown, BaseIssue<unknown>>,
+  ...args: unknown[]
+): UndefinedableSchema<
+  BaseSchema<unknown, unknown, BaseIssue<unknown>>,
+  unknown
+> {
+  // Create schema object
+  // @ts-expect-error
+  const schema: UndefinedableSchema<
+    BaseSchema<unknown, unknown, BaseIssue<unknown>>,
+    unknown
+  > = {
+    kind: 'schema',
+    type: 'undefinedable',
+    reference: undefinedable,
+    expects: `(${wrapped.expects} | undefined)`,
+    async: false,
+    wrapped,
+    _run(dataset, config) {
+      // If value is `undefined`, override it with default or return dataset
+      if (dataset.value === undefined) {
+        // If default is specified, override value of dataset
+        if ('default' in this) {
+          dataset.value = getDefault(
+            this,
+            dataset as Dataset<undefined, never>,
+            config
+          );
+        }
+
+        // If value is still `undefined`, return dataset
+        if (dataset.value === undefined) {
+          dataset.typed = true;
+          return dataset;
+        }
+      }
+
+      // Otherwise, return dataset of wrapped schema
+      return this.wrapped._run(dataset, config);
+    },
+  };
+
+  // Add default if specified
+  if (0 in args) {
+    // @ts-expect-error
+    schema.default = args[0];
+  }
+
+  // Return schema object
+  return schema;
+}