+++ website/src/routes/api/(types)/UndefinedableSchema/properties.ts
@@ -1,0 +1,141 @@
+import type { PropertyProps } from '~/components';
+
+export const properties: Record<string, PropertyProps> = {
+  TWrapped: {
+    modifier: 'extends',
+    type: {
+      type: 'custom',
+      name: 'BaseSchema',
+      href: '../BaseSchema/',
+      generics: [
+        'unknown',
+        'unknown',
+        {
+          type: 'custom',
+          name: 'BaseIssue',
+          href: '../BaseIssue/',
+          generics: ['unknown'],
+        },
+      ],
+    },
+  },
+  TDefault: {
+    modifier: 'extends',
+    type: {
+      type: 'custom',
+      name: 'Default',
+      href: '../Default/',
+      generics: [
+        {
+          type: 'custom',
+          name: 'TWrapped',
+        },
+        'undefined',
+      ],
+    },
+  },
+  BaseSchema: {
+    modifier: 'extends',
+    type: {
+      type: 'custom',
+      name: 'BaseSchema',
+      href: '../BaseSchema/',
+      generics: [
+        {
+          type: 'union',
+          options: [
+            {
+              type: 'custom',
+              name: 'InferInput',
+              href: '../InferInput/',
+              generics: [
+                {
+                  type: 'custom',
+                  name: 'TWrapped',
+                },
+              ],
+            },
+            'undefined',
+          ],
+        },
+        {
+          type: 'custom',
+          name: 'InferUndefinedableOutput',
+          href: '../InferUndefinedableOutput/',
+          generics: [
+            {
+              type: 'custom',
+              name: 'TWrapped',
+            },
+            {
+              type: 'custom',
+              name: 'TDefault',
+            },
+          ],
+        },
+        {
+          type: 'custom',
+          name: 'InferIssue',
+          href: '../InferIssue/',
+          generics: [
+            {
+              type: 'custom',
+              name: 'TWrapped',
+            },
+          ],
+        },
+      ],
+    },
+  },
+  type: {
+    type: {
+      type: 'string',
+      value: 'undefinedable',
+    },
+  },
+  reference: {
+    type: {
+      type: 'custom',
+      modifier: 'typeof',
+      name: 'undefinedable',
+      href: '../undefinedable/',
+    },
+  },
+  expects: {
+    type: {
+      type: 'template',
+      parts: [
+        {
+          type: 'string',
+          value: '(',
+        },
+        {
+          type: 'custom',
+          name: 'TWrapped',
+          indexes: [
+            {
+              type: 'string',
+              value: 'expects',
+            },
+          ],
+        },
+        {
+          type: 'string',
+          value: ' | undefined)',
+        },
+      ],
+    },
+  },
+  wrapped: {
+    type: {
+      type: 'custom',
+      name: 'TWrapped',
+    },
+  },
+  default: {
+    type: {
+      type: 'custom',
+      name: 'TDefault',
+    },
+  },
+};