+++ library/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Add `undefinedable` and `undefinedableAsync` schema (issue #385)
+
 ## v0.39.0 (August 24, 2024)
 
 - Add support for `exactOptionalPropertyTypes` config (issue #385)